+++ src/zlib/inflate.h
@@ -82,40 +82,40 @@ struct inflate_state {
     int havedict;               /* true if dictionary provided */
     int flags;                  /* gzip header method and flags (0 if zlib) */
     unsigned dmax;              /* zlib header max distance (INFLATE_STRICT) */
-    unsigned long check;        /* protected copy of check value */
+    uint32_t check;             /* protected copy of check value */
     unsigned long total;        /* protected copy of output count */
     gz_header* head;            /* where to save gzip header information */
         /* sliding window */
-    unsigned wbits;             /* log base 2 of requested window size */
-    unsigned wsize;             /* window size or zero if not using window */
-    unsigned whave;             /* valid bytes in the window */
-    unsigned wnext;             /* window write index */
+    uint16_t wbits;             /* log base 2 of requested window size */
+    uint64_t wsize;             /* window size or zero if not using window */
+    uint64_t whave;             /* valid bytes in the window */
+    uint64_t wnext;             /* window write index */
     unsigned char  *window;  /* allocated sliding window, if needed */
         /* bit accumulator */
     unsigned long hold;         /* input bit accumulator */
-    unsigned bits;              /* number of bits in "in" */
+    uint64_t bits;              /* number of bits in "in" */
         /* for string and stored block copying */
-    unsigned length;            /* literal or length of data to copy */
-    unsigned offset;            /* distance back to copy string from */
+    uint64_t length;            /* literal or length of data to copy */
+    uint64_t offset;            /* distance back to copy string from */
         /* for table and code decoding */
     unsigned extra;             /* extra bits needed */
         /* fixed and dynamic code tables */
     code const  *lencode;    /* starting table for length/literal codes */
     code const  *distcode;   /* starting table for distance codes */
-    unsigned lenbits;           /* index bits for lencode */
-    unsigned distbits;          /* index bits for distcode */
+    uint64_t lenbits;           /* index bits for lencode */
+    uint64_t distbits;          /* index bits for distcode */
         /* dynamic table building */
-    unsigned ncode;             /* number of code length code lengths */
-    unsigned nlen;              /* number of length code lengths */
-    unsigned ndist;             /* number of distance code lengths */
-    unsigned have;              /* number of code lengths in lens[] */
+    uint64_t ncode;             /* number of code length code lengths */
+    uint64_t nlen;              /* number of length code lengths */
+    uint64_t ndist;             /* number of distance code lengths */
+    uint64_t have;              /* number of code lengths in lens[] */
     code  *next;             /* next available space in codes[] */
     unsigned short lens[320];   /* temporary storage for code lengths */
     unsigned short work[288];   /* work area for code table building */
     code codes[ENOUGH];         /* space for code tables */
     int sane;                   /* if false, allow invalid distance too far */
     int back;                   /* bits back of last unprocessed length/lit */
-    unsigned was;               /* initial length of match */
+    uint64_t was;               /* initial length of match */
 };
 
 }
